Change files to cpp and hpp in order to call cpp bsp code
